backend/requirements.txt
fastapi
uvicorn
python-jose
passlib[argon2]
